requirements.txt
 psycopg2-binary
 faker
 fastapi
 uvicorn
 python-dotenv
 click
 pandas
 sqlalchemy
 pydantic